[add] page permission

+++ app/controllers/books_controller.rb
@@ -31,6 +31,12 @@ class BooksController < ApplicationController
 
     def destroy
         @book = Book.find(params[:id])
+
+        if @book.user.id != current_user.id
+            redirect_to books_path, status: 303
+            return
+        end
+
         if @book.destroy
             redirect_to books_path, status: 303, flash: {success: "destory book successfully"}
         else
@@ -43,10 +49,20 @@ class BooksController < ApplicationController
 
     def edit
         @book = Book.find(params[:id])
+        if @book.user.id != current_user.id
+            redirect_to books_path, status: 303
+            return
+        end
     end
 
     def update
         @book = Book.find(params[:id])
+
+        if @book.user.id != current_user.id
+            redirect_to books_path, status: 303
+            return
+        end
+
         if @book.update(book_params)
             redirect_to book_path(params[:id]), flash: {success: "update book successfully"}
         else
